Front-end: Added basic HTML structure for Todo App.
Developed the foundational HTML structure for a basic Todo application. This includes a main heading (<h1>) to display the application title, an input field (<input>) for users to enter tasks, and a submit button to add new tasks to the list. Additionally, an unordered list (<ul>) has been included to display the tasks dynamically. Proper meta tags for character encoding and viewport settings have been incorporated to ensure accessibility and responsiveness. This commit establishes the groundwork for further enhancements, including CSS for styling and JavaScript for interactive functionality. I also added two logos which will be used for the designing aspect of the project.